// apps/web/src/generated/types/measurements/GetMeasurements.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const getMeasurements200ValueTypeEnum = {
    NUMERIC: "NUMERIC",
    CATEGORICAL: "CATEGORICAL",
    TEXT: "TEXT"
} as const;

export type GetMeasurements200ValueTypeEnumKey = (typeof getMeasurements200ValueTypeEnum)[keyof typeof getMeasurements200ValueTypeEnum];

/**
 * @description Response for status 200
*/
export type GetMeasurements200 = {
    /**
     * @type string
    */
    id: string;
    /**
     * @type string
    */
    experimentId: string;
    /**
     * @type string
    */
    experimentName: string;
    /**
     * @type string
    */
    measurementDefinitionId: string;
    /**
     * @type string
    */
    definitionName: string;
    /**
     * @type string
    */
    valueType: GetMeasurements200ValueTypeEnumKey;
    /**
     * @type number,null
    */
    numericValue: number | null;
    /**
     * @type string,null
    */
    unit: string | null;
    /**
     * @type string,null
    */
    categoricalValue: string | null;
    /**
     * @type string,null
    */
    textValue: string | null;
    /**
     * @type string,null
    */
    notes: string | null;
    /**
     * @type string
    */
    recordedAt: string;
    /**
     * @type string,null
    */
    recordedById: string | null;
    /**
     * @type string,null
    */
    recordedByName: string | null;
}[];

export type GetMeasurementsQueryResponse = GetMeasurements200;

export type GetMeasurementsQuery = {
    Response: GetMeasurements200;
    Errors: any;
};